+++ src/commands/index.ts
@@ -1,1 +1,3 @@
-export * from './FrontMatter';
+export * from './Article';
+export * from './Settings';
+export * from './StatusBar';