+++ src/helpers/ArticleHelper.ts
@@ -1,0 +1,30 @@
+import * as vscode from 'vscode';
+import * as matter from "gray-matter";
+
+export class ArticleHelper {
+  
+  /**
+   * Get the contents of the current article
+   * 
+   * @param editor 
+   */
+  public static getFrontMatter(editor: vscode.TextEditor) {
+    const article = matter(editor.document.getText());
+    if (article && article.data) {
+      return article;
+    }
+    return null;
+  }
+
+  /**
+   * Store the new information in the file
+   * 
+   * @param editor 
+   * @param article 
+   */
+  public static async update(editor: vscode.TextEditor, article: matter.GrayMatterFile<string>) {
+    const newMarkdown = matter.stringify(article.content, article.data);
+    const nrOfLines = editor.document.lineCount as number;
+    await editor.edit(builder => builder.replace(new vscode.Range(new vscode.Position(0, 0), new vscode.Position(nrOfLines, 0)), newMarkdown));
+  }
+}