+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log
 
+## [0.0.7] - 2019-08-28
+
+- Added command to remap tags or categories in all posts: `Front Matter: Remap tag/category in all articles`
+
 ## [0.0.6] - 2019-08-28
 
 - Updated `package.json` file to include preview label
+- Added the status bar item to quickly view and update the draft status of an article
 
 ## [0.0.5] - 2019-08-27